All Courses
The StudentRecord object can now load in all of the student's enrolled courses, setting up the course objects. Each course object also gets a class session object that corresponds to the class the student took.

Note: I commented out a bunch of stuff in the dbConnector that was causing errors when it was used in the StudentRecord object. It hasn't seemed to have had a major impact on the app's visible performance, but I thought I'd mention it just in case something comes of it.